+++ packages/camera/camera_avfoundation/ios/camera_avfoundation/Sources/camera_avfoundation/include/camera_avfoundation/FLTCam_Test.h
@@ -3,15 +3,10 @@
 // found in the LICENSE file.
 
 #import "FLTCam.h"
+#import "FLTCaptureDevice.h"
+#import "FLTCapturePhotoOutput.h"
+#import "FLTDeviceOrientationProviding.h"
 #import "FLTSavePhotoDelegate.h"
-
-/// Determines the video dimensions (width and height) for a given capture device format.
-/// Used in tests to mock CMVideoFormatDescriptionGetDimensions.
-typedef CMVideoDimensions (^VideoDimensionsForFormat)(AVCaptureDeviceFormat *);
-
-/// Factory block returning an AVCaptureDevice.
-/// Used in tests to inject a device into FLTCam.
-typedef AVCaptureDevice * (^CaptureDeviceFactory)(void);
 
 @interface FLTImageStreamHandler : NSObject <FlutterStreamHandler>
 
@@ -33,7 +28,7 @@ typedef AVCaptureDevice * (^CaptureDeviceFactory)(void);
 @property(readonly, nonatomic) AVCaptureVideoDataOutput *captureVideoOutput;
 
 /// The output for photo capturing. Exposed setter for unit tests.
-@property(strong, nonatomic) AVCapturePhotoOutput *capturePhotoOutput;
+@property(strong, nonatomic) NSObject<FLTCapturePhotoOutput> *capturePhotoOutput;
 
 /// True when images from the camera are being streamed.
 @property(assign, nonatomic) BOOL isStreamingImages;
@@ -51,30 +46,6 @@ typedef AVCaptureDevice * (^CaptureDeviceFactory)(void);
 - (void)captureOutput:(AVCaptureOutput *)output
     didOutputSampleBuffer:(CMSampleBufferRef)sampleBuffer
            fromConnection:(AVCaptureConnection *)connection;
-
-/// Initializes a camera instance.
-/// Allows for injecting dependencies that are usually internal.
-- (instancetype)initWithCameraName:(NSString *)cameraName
-                     mediaSettings:(FCPPlatformMediaSettings *)mediaSettings
-            mediaSettingsAVWrapper:(FLTCamMediaSettingsAVWrapper *)mediaSettingsAVWrapper
-                       orientation:(UIDeviceOrientation)orientation
-               videoCaptureSession:(AVCaptureSession *)videoCaptureSession
-               audioCaptureSession:(AVCaptureSession *)audioCaptureSession
-               captureSessionQueue:(dispatch_queue_t)captureSessionQueue
-                             error:(NSError **)error;
-
-///  Initializes a camera instance.
-///  Allows for testing with specified resolution, audio preference, orientation,
-///  and direct access to capture sessions and blocks.
-- (instancetype)initWithMediaSettings:(FCPPlatformMediaSettings *)mediaSettings
-               mediaSettingsAVWrapper:(FLTCamMediaSettingsAVWrapper *)mediaSettingsAVWrapper
-                          orientation:(UIDeviceOrientation)orientation
-                  videoCaptureSession:(AVCaptureSession *)videoCaptureSession
-                  audioCaptureSession:(AVCaptureSession *)audioCaptureSession
-                  captureSessionQueue:(dispatch_queue_t)captureSessionQueue
-                 captureDeviceFactory:(CaptureDeviceFactory)captureDeviceFactory
-             videoDimensionsForFormat:(VideoDimensionsForFormat)videoDimensionsForFormat
-                                error:(NSError **)error;
 
 /// Start streaming images.
 - (void)startImageStreamWithMessenger:(NSObject<FlutterBinaryMessenger> *)messenger