+++ packages/pigeon/test/swift_generator_test.dart
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:pigeon/ast.dart';
-import 'package:pigeon/swift_generator.dart';
+import 'package:pigeon/src/ast.dart';
+import 'package:pigeon/src/swift/swift_generator.dart';
 import 'package:test/test.dart';
 
 import 'dart_generator_test.dart';
@@ -258,6 +258,27 @@ void main() {
     expect(code, contains('var aFloat64List: FlutterStandardTypedData? = nil'));
   });
 
+  test('gen pigeon error type', () {
+    final Root root = Root(apis: <Api>[], classes: <Class>[], enums: <Enum>[]);
+    final StringBuffer sink = StringBuffer();
+    const SwiftOptions swiftOptions = SwiftOptions();
+    const SwiftGenerator generator = SwiftGenerator();
+
+    generator.generate(
+      swiftOptions,
+      root,
+      sink,
+      dartPackageName: DEFAULT_PACKAGE_NAME,
+    );
+    final String code = sink.toString();
+    expect(code, contains('class PigeonError: Error'));
+    expect(code, contains('let code: String'));
+    expect(code, contains('let message: String?'));
+    expect(code, contains('let details: Sendable?'));
+    expect(code,
+        contains('init(code: String, message: String?, details: Sendable?)'));
+  });
+
   test('gen one flutter api', () {
     final Root root = Root(apis: <Api>[
       AstFlutterApi(name: 'Api', methods: <Method>[
@@ -401,7 +422,7 @@ void main() {
     final String code = sink.toString();
     expect(code,
         contains('completion: @escaping (Result<Void, PigeonError>) -> Void'));
-    expect(code, contains('completion(.success(Void()))'));
+    expect(code, contains('completion(.success(()))'));
     expect(code, isNot(contains('if (')));
   });