+++ packages/camera/camera_avfoundation/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.9.18+4
+
+* Refactors implementations to reduce usage of OCMock in internal testing.
+
+## 0.9.18+3
+
+* Refactors implementations to reduce usage of OCMock in internal testing.
+
+## 0.9.18+2
+
+* Refactors implementations to reduce usage of OCMock in internal testing.
+
+## 0.9.18+1
+
+* Refactors implementations to reduce usage of OCMock in internal testing.
+
 ## 0.9.18
 
 * Adds API support query for image streaming.